Updated typing and cleaned page
